SPARKC-283: Adds CassandraOption for Skipping Empty Columns
Previously unset columns had to be set to null since the Cassandra
Protocol <= V3 did not support unbound parameters in bound statements.
The newere version of the protocol allows for this, and this patch
enables the connector to take advtange of that behavior.

This is done through use of a new option container, CassandraOption.
This class holds an Option and the behavior that a None should have upon
being written. With this class a row can have varying behavior from
column to column and row to row.

If an Unset is used on a C* version prior to Protcol Version 3 a warning
will be logged that the Unset functionality cannot be used and the value
will be treated as a Delete instead.