SPARKC-283-UNSET: Refactor CassandraOption to Sealed Trait
Instead of the CassandraOption holding an Option we are instead mimicing
the Option class itself. The Option has 3 possible states, Value, Unset
and Null. An implicit conversion is present to transform to a standard
scala Option (Unset and Null -> None). This class can be used as a
default for reading out of Cassandra.